remove empty app duplication and magic code
