+++ frontend/src/lib/components/apps/editor/appUtils.ts
@@ -13,6 +13,8 @@ import {
 	ccomponents,
 	components,
 	getRecommendedDimensionsByComponent,
+	presets,
+	processDimension,
 	type AppComponent,
 	type BaseComponent,
 	type InitialAppComponent,
@@ -36,17 +38,18 @@ import { sendUserToast } from '$lib/toast'
 import { getNextId } from '$lib/components/flows/idUtils'
 import { enterpriseLicense } from '$lib/stores'
 import gridHelp from '../svelte-grid/utils/helper'
+import { DEFAULT_THEME } from './componentsPanel/themeUtils'
 
 type GridItemLocation =
 	| {
-		type: 'grid'
-		gridItemIndex: number
-	}
+			type: 'grid'
+			gridItemIndex: number
+	  }
 	| {
-		type: 'subgrid'
-		subgridItemIndex: number
-		subgridKey: string
-	}
+			type: 'subgrid'
+			subgridItemIndex: number
+			subgridKey: string
+	  }
 interface GridItemWithLocation {
 	location: GridItemLocation
 	item: GridItem
@@ -187,7 +190,7 @@ export function selectId(
 	selectedComponent: Writable<string[] | undefined>,
 	app: App
 ) {
-	; (document?.activeElement as HTMLElement)?.blur()
+	;(document?.activeElement as HTMLElement)?.blur()
 	if (e.shiftKey) {
 		selectedComponent.update((old) => {
 			if (old && old?.[0]) {
@@ -492,11 +495,11 @@ export function appComponentFromType<T extends keyof typeof components>(
 			xData:
 				type === 'plotlycomponentv2' || type === 'chartjscomponentv2'
 					? {
-						type: 'evalv2',
-						fieldType: 'array',
-						expr: '[1, 2, 3, 4]',
-						connections: []
-					}
+							type: 'evalv2',
+							fieldType: 'array',
+							expr: '[1, 2, 3, 4]',
+							connections: []
+						}
 					: undefined,
 			...(extra ?? {})
 		}
@@ -845,33 +848,33 @@ export type InitConfig<
 		| EvalAppInput
 		| EvalV2AppInput
 		| {
-			type: 'oneOf'
-			selected: string
-			configuration: Record<
-				string,
-				Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput>
-			>
-		}
+				type: 'oneOf'
+				selected: string
+				configuration: Record<
+					string,
+					Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput>
+				>
+		  }
 	>
 > = {
-		[Property in keyof T]: T[Property] extends StaticAppInput
+	[Property in keyof T]: T[Property] extends StaticAppInput
 		? T[Property]['value'] | undefined
 		: T[Property] extends { type: 'oneOf' }
-		? {
-			type: 'oneOf'
-			selected: keyof T[Property]['configuration']
-			configuration: {
-				[Choice in keyof T[Property]['configuration']]: {
-					[IT in keyof T[Property]['configuration'][Choice]]: T[Property]['configuration'][Choice][IT] extends StaticAppInput
-					? T[Property]['configuration'][Choice][IT] extends StaticAppInputOnDemand
-					? () => Promise<T[Property]['configuration'][Choice][IT]['value'] | undefined>
-					: T[Property]['configuration'][Choice][IT]['value'] | undefined
-					: undefined
+			? {
+					type: 'oneOf'
+					selected: keyof T[Property]['configuration']
+					configuration: {
+						[Choice in keyof T[Property]['configuration']]: {
+							[IT in keyof T[Property]['configuration'][Choice]]: T[Property]['configuration'][Choice][IT] extends StaticAppInput
+								? T[Property]['configuration'][Choice][IT] extends StaticAppInputOnDemand
+									? () => Promise<T[Property]['configuration'][Choice][IT]['value'] | undefined>
+									: T[Property]['configuration'][Choice][IT]['value'] | undefined
+								: undefined
+						}
+					}
 				}
-			}
-		}
-		: undefined
-	}
+			: undefined
+}
 
 export function initConfig<
 	T extends Record<
@@ -880,13 +883,13 @@ export function initConfig<
 		| EvalAppInput
 		| EvalV2AppInput
 		| {
-			type: 'oneOf'
-			selected: string
-			configuration: Record<
-				string,
-				Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput>
-			>
-		}
+				type: 'oneOf'
+				selected: string
+				configuration: Record<
+					string,
+					Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput>
+				>
+		  }
 	>
 >(
 	r: T,
@@ -894,13 +897,13 @@ export function initConfig<
 		string,
 		| StaticAppInput
 		| {
-			type: 'oneOf'
-			selected: string
-			configuration: Record<
-				string,
-				Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput | boolean>
-			>
-		}
+				type: 'oneOf'
+				selected: string
+				configuration: Record<
+					string,
+					Record<string, StaticAppInput | EvalAppInput | EvalV2AppInput | boolean>
+				>
+		  }
 		| any
 	>
 ): InitConfig<T> {
@@ -910,31 +913,31 @@ export function initConfig<
 				Object.entries(r).map(([key, value]) =>
 					value.type == 'static'
 						? [
-							key,
-							configuration?.[key]?.type == 'static' ? configuration?.[key]?.['value'] : undefined
-						]
+								key,
+								configuration?.[key]?.type == 'static' ? configuration?.[key]?.['value'] : undefined
+							]
 						: value.type == 'oneOf'
 							? [
-								key,
-								{
-									selected: value.selected,
-									type: 'oneOf',
-									configuration: Object.fromEntries(
-										Object.entries(value.configuration).map(([choice, config]) => {
-											const conf = initConfig(
-												config,
-												configuration?.[key]?.configuration?.[choice]
-											)
-											Object.entries(config).forEach(([innerKey, innerValue]) => {
-												if (innerValue.type === 'static' && !(innerKey in conf)) {
-													conf[innerKey] = innerValue.value
-												}
+									key,
+									{
+										selected: value.selected,
+										type: 'oneOf',
+										configuration: Object.fromEntries(
+											Object.entries(value.configuration).map(([choice, config]) => {
+												const conf = initConfig(
+													config,
+													configuration?.[key]?.configuration?.[choice]
+												)
+												Object.entries(config).forEach(([innerKey, innerValue]) => {
+													if (innerValue.type === 'static' && !(innerKey in conf)) {
+														conf[innerKey] = innerValue.value
+													}
+												})
+												return [choice, conf]
 											})
-											return [choice, conf]
-										})
-									)
-								}
-							]
+										)
+									}
+								]
 							: [key, undefined]
 				)
 			) as any
@@ -1394,4 +1397,47 @@ export function animateTo(start: number, end: number, onUpdate: (newValue: numbe
 
 function easeInOut(t: number) {
 	return t < 0.5 ? 2 * t * t : -1 + (4 - 2 * t) * t
+}
+
+export function emptyApp(): App {
+	let value: App = {
+		grid: [],
+		fullscreen: false,
+		unusedInlineScripts: [],
+		hiddenInlineScripts: [],
+		theme: {
+			type: 'path',
+			path: DEFAULT_THEME
+		}
+	}
+	const preset = presets['topbarcomponent']
+
+	const id = insertNewGridItem(
+		value,
+		appComponentFromType(preset.targetComponent, preset.configuration, undefined, {
+			customCss: {
+				container: {
+					class: '!p-0' as any,
+					style: ''
+				}
+			}
+		}) as (id: string) => AppComponent,
+		undefined,
+		undefined,
+		'topbar',
+		{ x: 0, y: 0 },
+		{
+			3: processDimension(preset.dims, 3),
+			12: processDimension(preset.dims, 12)
+		},
+		true,
+		true
+	)
+
+	setUpTopBarComponentContent(id, value)
+
+	value.hideLegacyTopBar = true
+	value.mobileViewOnSmallerScreens = false
+	console.log('Empty app created', value)
+	return value
 }
